bat: Rename base to base_bat
To be consistent with the image BAT tests the package that contains the
basic BAT tests has been renamed from base to base_bat.
