+++ _release/bat/base_bat/base_bat_test.go
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-package base
+package base_bat
 
 import (
 	"context"
